# server/api-docs/openapi.yaml
swagger: "2.0"
info:
  description: "HTTP API for the Hilbert UI server."
  version: "1.0.0"
  title: "Hilbert UI"
  contact:
    email: "eric.londaits@imaginary.org"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api"
schemes:
- "http"
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /test-backend/flags:
    get:
      tags:
        - "Test backend"
      summary: "Get the test backend flags"
      responses:
        200:
          description: Returns the values of the flags
          schema:
            type: object
            properties:
              testMode:
                type: boolean
              flags:
                $ref: "#/definitions/TestBackendFlags"

    post:
      tags:
        - "Test backend"
      summary: "Set the test backend flags"
      parameters:
        - in: body
          name: flags
          description: The flag values you want to set
          schema:
            $ref: "#/definitions/TestBackendFlags"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/TestBackendFlags"
  /test-backend/stations/unreachable:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations becoming unreachable
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
        400:
          description: Missing 'ids' argument
  /test-backend/stations/reachable:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations becoming reachable
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
        400:
          description: Missing 'ids' argument
  /test-backend/stations/stop-unexpectedly:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations stopping unexpectedly
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
        400:
          description: Missing 'ids' argument
definitions:
...
  TestBackendFlags:
    type: object
    properties:
      sim-fail-cli:
        type: boolean
        default: false
      sim-unreachable:
        type: boolean
        default: false
      sim-timeout:
        type: boolean
        default: false